Tikhonov/lab5/requirements.txt
requests
SQLAlchemy
psycopg2-binary
kafka-python-ng
transformers
torch
sentence-transformers
numpy